Revert "[DHCP relay]: Add patch to always undef VLAN_TCI_PRESENT so as not to treat VLAN-tagged packets differently (#1254)" (#1291)
This reverts commit 64602ad5b2294f2e1de6b0f1f1761fb7697b21a7.

Reverting this commit because the root cause of the issue was due to
a Broadcom SDK bug which caused VLAN-tagged packets to be tagged twice.
When packets were double-tagged, only one tag was getting stripped before
the packet was trapped to the CPU. However, the second tag was left behind,
which would cause the DHCP relay agent to ignore the packet. The bug has
since been fixed, so now all VLAN-tagged packets should properly have their
tag removed before being trapped to the CPU. Therefore the DHCP relay
agent should never encounter a VLAN-tagged packet.