[Mellanox] fix sfp issue in reading qsfp_status (#3386)
There is a very low possibility that qsfp_status exists when opening it but doesn't when reading it a bit later. If this occurs an exception is raised.
However, this should be treated as "not presence" rather than causing an exception, which can be achieved by putting it into try block.