[Juniper][QFX5210] Workaround for orchagent crash (#3458)
Orchagent is crashing on the latest SONiC images. Issue #458
is raised to track the problem. Dynamic port breakout
commit 6f40933d3d7b9f21a97de275fbd14ea3598d9a0a introduced
this regression. One of the recommendation that we received
was to disable the loopback and mgmt ports in bcm config.
It was helpful in fixing the issue but not the ideal solution.
It works on qfx5210 as the loopback and management configurations
are not applicable.

While the orchagent crash is being debugged, we need this commit
to fix the issue in qfx5210 platform. This patch can be reversed
once the correct fix for the orchagent is identified.

Signed-off-by: Ciju Rajan K <crajank@juniper.net>