[lldpmgrd] only log error in is_port_up() after port init done (#4606)
lldpmgrd listens for changes to the PORT table in the CONFIG_DB and APP_DB in order to handle alias/description config change. It checks if port is up or down by looking into the oper-status for in APP_DB PORT TABLE. If it cannot find it in the App DB, it will log error. 

During initializing, it is possible that there is a port change in CONFIG_DB and but the not ready in APP_DB. 

The change here is to only log error in is_port_up() after port init done.