[build] Increase size of dockerfs ramdisk to accomodate more containers (#4666)
Images built from master branch and installed on devices where we mount /var/lib/docker in RAM (because the HDD is small) were failing to boot properly. The Docker service failed to start because /var/lib/docker was filled to 100%. This is due to the increase in total number of containers in the image.

As of today, /var/lib/docker contains 1.3 GB of data. Therefore, this PR increases the size of the ramdisk to 1.5 GB to accommodate all the containers. Example output below from an Arista-7050-QX32 SKU:

```
admin@sonic:~$ df -h
Filesystem      Size  Used Avail Use% Mounted on
...
tmpfs           1.5G  1.3G  172M  89% /var/lib/docker
...
```