[sonic-utilities] Add Python 3 sonic-yang-mgmt package as a dependency (#5903)
Add Python 3 sonic-yang-mgmt package as a dependency in sonic-utilities makefile.

Recently updated the sonic-utilities Jenkins build environment [here](https://github.com/Azure/sonic-build-tools/pull/185) to begin running tests on the Python 3 version of the sonic-utilities package. However, the build is failing because it attempts to copy the Python 3 sonic-yang-mgmt package from the artifacts of the latest VS build, but the package does not exist because there are no targets which specify it as a dependency. This PR will ensure the Python 3 package is built during the image build process.